* commit '442ec059250da306692c1f22dc468dd9e3fb1bd7':
  Remove debugging code
  Update MPEG video implementation to support large files
  Update Matroska implementation to support large files
  Remove x86 specific compiler flag
  Update MP4 implementation to support large files and atoms
  Change File base class to support large files
  Change IOStream interface and implementations to support large files
  Enabe LFS support
  Force build 32-bit target even on 64-bit hosts

+++ taglib/matroska/ebmlreader.h
@@ -9,14 +9,14 @@ namespace TagLib {
     class EBMLReader
     {
     public:
-      EBMLReader(File *_file, ulong position);
-      EBMLReader(const EBMLReader &parent, ulong position, uint id = 0, ulong size = 0);
+      EBMLReader(File *_file, long long position);
+      EBMLReader(const EBMLReader &parent, long long position, uint id = 0, long long size = 0);
 
       MatroskaID id() const;
-      ulong size() const;
-      ulong getDataSize() const;
-      ulong getDataOffset() const;
-      ulong getOffset() const;
+      long long size() const;
+      long long getDataSize() const;
+      long long getDataOffset() const;
+      long long getOffset() const;
       const EBMLReader *getParent() const;
 
       bool isValid() const;
@@ -31,11 +31,11 @@ namespace TagLib {
       File* file;
       const EBMLReader *parent;
 
-      ulong offset;
+      long long offset;
       uint ebmlId;
 
-      ulong dataOffset;
-      ulong dataSize;
+      long long dataOffset;
+      long long dataSize;
 
       bool isSuccessRead;
